experiment with refactoring tests
